+++ ApplicationCore/Model/InnerCore/InnerCoreModData.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace ICModsFunctions
+namespace ApplicationCore.Model.InnerCore
 {
     public class InnerCoreModData
     {
