+++ ICModsFunctions.sln
@@ -5,6 +5,10 @@ VisualStudioVersion = 17.0.31710.8
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ICModsFunctions", "ICModsFunctions\ICModsFunctions.csproj", "{958D441B-DC51-4FFA-AFFE-8C0CD446D9D3}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "Infrastructure\Infrastructure.csproj", "{0B3DA823-EAA7-48F6-9C30-C4D23891A423}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApplicationCore", "ApplicationCore\ApplicationCore.csproj", "{5880CFE6-EA28-4347-9DE8-99D98AC5FF83}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +19,14 @@ Global
 		{958D441B-DC51-4FFA-AFFE-8C0CD446D9D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{958D441B-DC51-4FFA-AFFE-8C0CD446D9D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{958D441B-DC51-4FFA-AFFE-8C0CD446D9D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0B3DA823-EAA7-48F6-9C30-C4D23891A423}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0B3DA823-EAA7-48F6-9C30-C4D23891A423}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0B3DA823-EAA7-48F6-9C30-C4D23891A423}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0B3DA823-EAA7-48F6-9C30-C4D23891A423}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5880CFE6-EA28-4347-9DE8-99D98AC5FF83}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5880CFE6-EA28-4347-9DE8-99D98AC5FF83}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5880CFE6-EA28-4347-9DE8-99D98AC5FF83}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5880CFE6-EA28-4347-9DE8-99D98AC5FF83}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
