+++ ICModsFunctions/ManualUpdateCosmosModStat.cs
@@ -2,7 +2,6 @@ using System;
 using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
-using ICModsFunctions.Model.Cosmos;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.WebJobs;
@@ -13,6 +12,9 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
 using Newtonsoft.Json;
+using ApplicationCore;
+using ApplicationCore.Model.Cosmos;
+using Infrastructure.Data.Cosmos;
 
 namespace ICModsFunctions
 {
