+++ ICModsFunctions/UpdateCosmosStats.cs
@@ -4,7 +4,9 @@ using System.Linq;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using ICModsFunctions.Model.Cosmos;
+using ApplicationCore.Model.Cosmos;
+using ApplicationCore.Model.InnerCore;
+using Infrastructure.Data.Cosmos;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using Microsoft.Azure.WebJobs.Extensions.Http;
